Moving PodioOutput from Examples to FWCore
