+++ Examples/CMakeLists.txt
@@ -6,9 +6,7 @@ gaudi_subdir(Examples v1r0)
 # this declaration will not be needed in the future
 gaudi_depends_on_subdirs(GaudiAlg DataObjects FWCore)
 
-find_package(ROOT COMPONENTS RIO Tree)
-
 gaudi_add_module(Examples
                  src/*.cpp
-                 LINK_LIBRARIES GaudiAlgLib DataObjects FWCore ROOT)
+                 LINK_LIBRARIES GaudiAlgLib DataObjects FWCore)
