+++ FWCore/CMakeLists.txt
@@ -5,6 +5,7 @@ gaudi_subdir(FWCore v1r0)
 
 find_package(FCCEDM)
 find_package(PODIO)
+find_package(ROOT COMPONENTS RIO Tree)
 
 # this declaration will not be needed in the future
 gaudi_depends_on_subdirs(GaudiAlg GaudiKernel)
@@ -17,4 +18,4 @@ gaudi_add_library(FWCore
 
 gaudi_add_module(FWCorePlugins
                  components/*.cpp
-                 LINK_LIBRARIES GaudiAlgLib GaudiKernel FWCore)
+                 LINK_LIBRARIES GaudiAlgLib GaudiKernel FWCore ROOT)
